move label out of nested directory structure
