+++ frontend/src/component/common/NewConstraintAccordion/LegalValueLabel/LegalValueLabel.styles.ts
