+++ frontend/src/component/common/NewConstraintAccordion/LegalValueLabel/LegalValueLabel.tsx
